Refactor selection style strategies into separate modules
- Moved CheckboxStrategy, RadioStrategy, ChipStrategy, and SwitchStrategy implementations into their respective files for better organization and maintainability.
- Updated the main strategy module to import these implementations.
- Enhanced the documentation for each strategy to clarify their purpose and behavior according to Material Design 3 specifications.
